backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
python-decouple>=3.8
firebase-admin>=6.2.0
httpx>=0.24.1
pytest>=7.3.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sentry-sdk>=1.28.0
stripe>=5.4.0
mapbox>=0.18.1
loguru>=0.7.0
strawberry-graphql>=0.198.0 